requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LLM & Agent
langchain-openai>=0.3
langgraph>=0.2

# Utilities
python-multipart>=0.0.6